+++ internal/seer/seer.go
