+++ internal/seer/seer_test.go
